refactor: replaced native code with the one from react-native-compat, added compat as peer dep
