+++ package.json
@@ -32,6 +32,7 @@
     "@types/qrcode": "1.5.1",
     "@types/react": "18.2.34",
     "@types/react-native": "0.72.2",
+    "@walletconnect/react-native-compat": "2.10.5",
     "eslint": "^8.46.0",
     "eslint-plugin-ft-flow": "2.0.3",
     "eslint-plugin-prettier": "5.0.1",
