--- packages/core/android/src/main/AndroidManifestNew.xml
@@ -1,2 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-</manifest>
