--- packages/core/android/src/oldarch/InstalledAppSpec.kt
@@ -1,11 +1,0 @@
-package com.installedapp
-
-import com.facebook.react.bridge.ReactApplicationContext
-import com.facebook.react.bridge.ReactContextBaseJavaModule
-import com.facebook.react.bridge.Promise
-
-abstract class InstalledAppSpec internal constructor(context: ReactApplicationContext) :
-  ReactContextBaseJavaModule(context) {
-
-  abstract fun isAppInstalled(packageName: String?, promise: Promise)
-}
