+++ packages/core/package.json
@@ -13,18 +13,7 @@
   },
   "files": [
     "src",
-    "lib",
-    "android",
-    "!android/build",
-    "!android/gradle",
-    "!android/gradlew",
-    "!android/gradlew.bat",
-    "!android/local.properties",
-    "!ios/build",
-    "!**/__tests__",
-    "!**/__fixtures__",
-    "!**/__mocks__",
-    "!**/.*"
+    "lib"
   ],
   "keywords": [
     "web3",
@@ -50,6 +39,7 @@
   },
   "peerDependencies": {
     "@react-native-async-storage/async-storage": ">=1.17.0",
+    "@walletconnect/react-native-compat": ">=2.10.5",
     "react": ">=17",
     "react-native": ">=0.68.5"
   },
